Added bid
Added possibility to make a bid on the auction items

+++ shared/src/commonMain/kotlin/com/example/kmm/model/Product.kt
@@ -6,7 +6,7 @@ data class Product(
     val name: String,
     val make: String? = null,
     val description: String? = null,
-    val currentBid: Int,
+    val currentBid: Double,
     val endDate: String,
     val reservePriceStatus: String,
     val municipalityName: String,
